Rename module in code.
