+++ xcp_d/workflows/__init__.py
@@ -2,7 +2,7 @@
 # vi: set ft=python sts=4 ts=4 sw=4 et
 """Nipype workflows for xcp_d."""
 
-from xcp_d.workflow import (
+from xcp_d.workflows import (
     anatomical,
     base,
     bold,
