+++ xcp_d/workflows/outputs.py
@@ -31,7 +31,7 @@ def init_writederivatives_wf(
             :graph2use: orig
             :simple_form: yes
 
-            from xcp_d.workflow.outputs import init_writederivatives_wf
+            from xcp_d.workflows.outputs import init_writederivatives_wf
             wf = init_writederivatives_wf(
                 bold_file="/path/to/file.nii.gz",
                 bandpass_filter=True,
